This push reintroduces the form complete with validation, and includes a small change to the host's view:
Now when the host is viewing an event they are hosting and they have no
attendees, they will not be presented with "Here are your attendees!"
