+++ client/modules/Event/components/EventPageDetails/Host/Attendees/Attendees.js
@@ -3,7 +3,7 @@ import React from 'react';
 /* eslint-disable react/prop-types */
 const Attendees = ({ attendees }) => (
   <div>
-    These users are currently subscribed to your event!
+      {attendees.length ? 'These users are currently subscribed to your event!' : ''}
     {
       attendees.map((user) => {
         return (
